+++ posthog-react-native/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next
+
+1. Added new `const [flag, payload] = useFeatureFlagWithPayload('my-flag-name')` hook that returns the flag result and it's payload if it has one.
+
 # 2.7.1 - 2023-05-25
 
 1. The `$screen_name` property will be registered for all events whenever `screen` is called